fix(core): move the shared commands out of the commands namespace
